BBOX CRS/SRID fix for sql provider
This address and issue where CRS to_epsg returns null. This causes the envelop to use the default SRID, which was not necessary compatible with the table. This commit changes that default to the geometry column of the table; assuming the user has correctly configured the provider to ensure the coordinates are compatible.